Mac64 build fix.
reinterpret_cast from void* to int causes compile error in 64bit environment.


BUG=None
TEST=None


Review URL: http://codereview.chromium.org/9494009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@124114 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/system_monitor/system_monitor_mac.mm
@@ -1,4 +1,4 @@
-// Copyright (c) 2011 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -28,7 +28,7 @@ void SystemPowerEventCallback(void*,
     case kIOMessageSystemWillSleep:
       sys_monitor->ProcessPowerMessage(SystemMonitor::SUSPEND_EVENT);
       IOAllowPowerChange(g_system_power_io_port,
-          reinterpret_cast<int>(message_argument));
+          reinterpret_cast<intptr_t>(message_argument));
       break;
 
     case kIOMessageSystemWillPowerOn:
